server/requirements.txt
Django==5.2.2
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.6.0
django-cloudinary-storage
cloudinary==1.36.0
python-dotenv==1.0.1
bcrypt==4.1.3
matplotlib
pandas
numpy
google-auth
google-auth-httplib2
PyJWT
channels==4.0.0
daphne==4.1.0
# channels['daphne']
whitenoise
django-ipware
mysqlclient